SCRD-1252 Fix input-model-ref README.html files
During the rebranding process we noticed that some of the README.html
files in the input-model-ref repo were missing footers or had <small>
sections without corresponding closing </small> for Copyright lines.

This patch addresses both issues for the examples but not the 4.0/CG
README.html files, which are incomplete placeholders.

Change-Id: I91814577b47e77ef25484fca5a0edabfbc608eb4